ET-271 As a student, I want my learner progress to be reset when I re-take a course via LTI
